requirements.txt
astroid==1.5.3
click==6.7
colorama==0.3.9
decorator==4.1.2
Flask==0.12.2
Flask-SQLAlchemy==2.3.2
isort==4.2.15
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
pbr==3.1.1
pylint==1.7.4
six==1.11.0
SQLAlchemy==1.1.15
sqlalchemy-migrate==0.11.0
sqlparse==0.2.4
Tempita==0.5.2
Werkzeug==0.12.2
wrapt==1.10.11
